update orm
update table finance-vehicle-type
Interest to float64,
table add user 4 to 8 column

+++ orm/finance-vehicle-type.go
@@ -6,9 +6,9 @@ type FinanceVehicleType struct {
 
 	DomainCode string ` db:"domain_code" json:"domain_code" gorm:"type:varchar(50);index " `
 
-	FinanceID     string ` db:"finance_id" json:"finance_id" gorm:"type:varchar(36); " `
-	VehicleTypeID string ` db:"vehicle_type_id" json:"vehicle_type_id" gorm:"type:varchar(36);  " `
-	Month         int    ` db:"month" json:"month" gorm:"type:varchar(50);  " `
-	Interest      int    ` db:"interest" json:"interest" gorm:"type:float; " `
-	RowOrder      int    ` db:"row_order" json:"row_order" gorm:"type:int(11); " `
+	FinanceID     string  ` db:"finance_id" json:"finance_id" gorm:"type:varchar(36); " `
+	VehicleTypeID string  ` db:"vehicle_type_id" json:"vehicle_type_id" gorm:"type:varchar(36);  " `
+	Month         int     ` db:"month" json:"month" gorm:"type:varchar(50);  " `
+	Interest      float64 ` db:"interest" json:"interest" gorm:"type:float; " `
+	RowOrder      int     ` db:"row_order" json:"row_order" gorm:"type:int(11); " `
 }
